+++ design-system/heading/src/heading.tsx
@@ -29,67 +29,67 @@ const headingResetStyles = css({
 });
 
 const h900Styles = css({
-  fontSize: '35px',
-  fontWeight: 500,
+  fontSize: token('font.size.600', '35px'),
+  fontWeight: token('font.weight.medium', '500'),
   letterSpacing: '-0.01em',
-  lineHeight: '40px',
+  lineHeight: token('font.lineHeight.600', '40px'),
 });
 
 const h800Styles = css({
-  fontSize: '29px',
-  fontWeight: 600,
+  fontSize: token('font.size.500', '29px'),
+  fontWeight: token('font.weight.semibold', '600'),
   letterSpacing: '-0.01em',
-  lineHeight: '32px',
+  lineHeight: token('font.lineHeight.500', '32px'),
 });
 
 const h700Styles = css({
-  fontSize: 24,
-  fontWeight: 500,
+  fontSize: token('font.size.400', '24px'),
+  fontWeight: token('font.weight.medium', '500'),
   letterSpacing: '-0.01em',
-  lineHeight: '28px',
+  lineHeight: token('font.lineHeight.400', '28px'),
 });
 
 const h600Styles = css({
-  fontSize: 20,
-  fontWeight: 500,
+  fontSize: token('font.size.300', '20px'),
+  fontWeight: token('font.weight.medium', '500'),
   letterSpacing: '-0.008em',
-  lineHeight: '24px',
+  lineHeight: token('font.lineHeight.300', '24px'),
 });
 
 const h500Styles = css({
-  fontSize: 16,
-  fontWeight: 600,
+  fontSize: token('font.size.200', '16px'),
+  fontWeight: token('font.weight.semibold', '600'),
   letterSpacing: '-0.006em',
-  lineHeight: '20px',
+  lineHeight: token('font.lineHeight.200', '20px'),
 });
 
 const h400Styles = css({
-  fontSize: 14,
-  fontWeight: 600,
+  fontSize: token('font.size.100', '14px'),
+  fontWeight: token('font.weight.semibold', '600'),
   letterSpacing: '-0.003em',
-  lineHeight: '16px',
+  lineHeight: token('font.lineHeight.100', '16px'),
 });
 
 const h300Styles = css({
-  fontSize: 12,
-  fontWeight: 600,
+  fontSize: token('font.size.075', '12px'),
+  fontWeight: token('font.weight.semibold', '600'),
   letterSpacing: 0,
-  lineHeight: '16px',
+  lineHeight: token('font.lineHeight.100', '16px'),
   textTransform: 'uppercase',
 });
 
 const h200Styles = css({
-  fontSize: 12,
-  fontWeight: 600,
+  fontSize: token('font.size.075', '12px'),
+  fontWeight: token('font.weight.semibold', '600'),
   letterSpacing: 0,
-  lineHeight: '16px',
+  lineHeight: token('font.lineHeight.100', '16px'),
 });
 
 const h100Styles = css({
-  fontSize: 11,
-  fontWeight: 700,
+  fontSize: token('font.size.050', '11px'),
+  fontWeight: token('font.weight.bold', '700'),
   letterSpacing: 0,
-  lineHeight: '16px',
+  lineHeight: token('font.lineHeight.100', '16px'),
 });
 
 const inverseStyles = css({