+++ linking-platform/smart-card/src/view/HoverCard/components/views/unauthorised/__tests__/index.test.tsx
@@ -89,7 +89,7 @@ describe('Unauthorised Hover Card', () => {
     expect(iconElement).toBeTruthy();
     expect(titleElement.textContent).toBe('Connect your Google account');
     expect(mainContentElement.textContent).toBe(
-      'Connect Google to Atlassian to view more details from your work and collaboration from one place. Learn more about Smart Links.',
+      'Connect Google to Atlassian to view more details of your work and collaborate from one place. Learn more about Smart Links.',
     );
     expect(buttonElement.textContent).toBe('Connect to Google');
   });