+++ linking-platform/smart-card/src/view/__tests__/card/card-states.flexible-card.test.tsx
@@ -455,7 +455,7 @@ describe('smart-card: card states, flexible block withUrl', () => {
       );
       expect(unauthorizedContent).toBeTruthy();
       expect(unauthorizedContent.textContent).toBe(
-        'Connect to Atlassian to view more details from your work and collaboration from one place. Learn more about Smart Links.',
+        'Connect to Atlassian to view more details of your work and collaborate from one place. Learn more about Smart Links.',
       );
 
       const providerImage = queryByTestId(
@@ -538,7 +538,7 @@ describe('smart-card: card states, flexible block withUrl', () => {
       );
       expect(unauthorizedContent).toBeTruthy();
       expect(unauthorizedContent.textContent).toBe(
-        'Connect Google to Atlassian to view more details from your work and collaboration from one place. Learn more about Smart Links.',
+        'Connect Google to Atlassian to view more details of your work and collaborate from one place. Learn more about Smart Links.',
       );
 
       const providerImage = getByTestId(