+++ platform/feature-flags/src/index.ts
@@ -1,4 +1,5 @@
 import { store } from './registry';
+import { BooleanFeatureFlagType } from './types';
 
 /**
  * Sets a feature flag to a given value.
@@ -6,7 +7,10 @@ import { store } from './registry';
  * @param name
  * @param value
  */
-export function setBooleanFF(name: string, value: boolean): void {
+export function setBooleanFF(
+  name: BooleanFeatureFlagType,
+  value: boolean,
+): void {
   store[name] = value;
 }
 
@@ -14,7 +18,7 @@ export function setBooleanFF(name: string, value: boolean): void {
  * Returns the value of a feature flag. If the flag is not set, returns the "false" as a default value.
  * @param name
  */
-export function getBooleanFF(name: string): boolean {
+export function getBooleanFF(name: BooleanFeatureFlagType): boolean {
   const defaultValue = false;
 
   return store?.[name] ?? defaultValue;